+++ src/App.tsx
@@ -2,6 +2,7 @@ import { Routes, Route } from 'react-router-dom'
 import Layout from './components/Layout/Layout'
 import Home from './pages/Home'
 import DiagramPage from './pages/DiagramPage'
+import EmbedPage from './pages/EmbedPage'
 
 export default function App() {
   return (
@@ -10,6 +11,7 @@ export default function App() {
         <Route index element={<Home />} />
         <Route path="diagram/:widgetId" element={<DiagramPage />} />
       </Route>
+      <Route path="embed/:widgetId" element={<EmbedPage />} />
     </Routes>
   )
 }